requirements.txt
amqp==2.3.2
billiard==3.5.0.3
boto3==1.7.67
celery==4.1.1
certifi==2018.4.16
chardet==3.0.4
clickhouse-driver==0.0.8
dj-database-url==0.4.1
Django==1.11.1
django-celery-beat==1.1.1
django-celery-results==1.0.1
django-daterange-filter==1.0.0
django-debug-toolbar==1.9.1
django-extensions==2.0.7
django-filter==1.0.4
django-global-permissions==0.2.3
django-storages==1.6.6
djangorestframework==3.6.4
djangorestframework-csv==2.1.0
django-dbbackup==3.2.0
gunicorn==19.6.0
idna==2.6
kombu==4.2.1
Markdown==2.6.9
psycopg2==2.7.4
pytz==2018.4
redis==2.10.6
requests==2.18.4
six==1.11.0
sortedcontainers==1.5.9
sqlparse==0.2.4
unicodecsv==0.14.1
urllib3==1.22
vine==1.1.4
whitenoise==3.2
openpyxl==2.5.9